+++ src/components/modules/Learn/index.js
@@ -1,0 +1,1 @@
+export { default } from ".Learn.js";
